Moved helper classes into helper package
