+++ helpers/sqlalchemy_helpers.py
@@ -1,0 +1,79 @@
+from dataclasses import dataclass
+from dataclasses import field
+from typing import Any
+from typing import Dict
+from typing import List
+from typing import Optional
+from typing import Union
+
+from sqlalchemy import Column
+from sqlalchemy.sql import FromClause
+from sqlalchemy.sql import Selectable
+from sqlalchemy.sql.elements import UnaryExpression
+
+from typeshed import JoinListType
+from typeshed import JoinStruct
+from typeshed import SQLLogicType
+
+
+@dataclass
+class QueryHandler:
+    func_str: str
+    func_data: Any
+    allow_empty_data: bool = field(default=False)
+
+    def update_query(self, query: Selectable) -> Selectable:
+        if self.func_data or self.allow_empty_data and self.func_data is not None:
+            query = getattr(query, self.func_str)(*self.func_data)
+        return query
+
+
+class DictQueryHandler(QueryHandler):
+    func_data: Optional[dict]
+
+    def update_query(self, query: Selectable) -> Selectable:
+        if self.func_data:
+            query = getattr(query, self.func_str)(**self.func_data)
+        return query
+
+
+class JoinQueryHandler(QueryHandler):
+    func_data: Optional[JoinListType]
+
+    def update_query(self, query: FromClause) -> Selectable:  # type: ignore[override]
+        if self.func_data:
+            for join_on in self.func_data:
+                if isinstance(join_on, tuple):
+                    query = query.join(*join_on)
+                elif isinstance(join_on, JoinStruct):
+                    join_data = join_on.get_join_data()
+                    join_func = join_on.get_join_func()
+                    query = getattr(query, join_func)(*join_data)
+                else:
+                    query = query.join(join_on)
+        return query
+
+
+@dataclass
+class QueryArgs:  # pylint: disable=R0902
+    filter_list: Optional[List[SQLLogicType]] = None
+    filter_dict: Optional[Dict[str, Any]] = None
+    eager_options: Optional[List] = None
+    order_by_list: Optional[List[Union[Column, UnaryExpression]]] = None
+    join_list: Optional[JoinListType] = None
+    distinct_on_list: Optional[List[Optional[Column]]] = None
+    group_by_list: Optional[List[Column]] = None
+    having_list: Optional[List[SQLLogicType]] = None
+
+    def get_query_handlers(self) -> List[QueryHandler]:
+        query_handlers = [
+            DictQueryHandler("filter_by", self.filter_dict),
+            JoinQueryHandler("join", self.join_list),
+            QueryHandler("filter", self.filter_list),
+            QueryHandler("group_by", self.group_by_list),
+            QueryHandler("having", self.having_list),
+            QueryHandler("options", self.eager_options),
+            QueryHandler("order_by", self.order_by_list),
+            QueryHandler("distinct", self.distinct_on_list, allow_empty_data=True),
+        ]
+        return query_handlers
